+++ src/Shared/Infrastructure/LLMChain/EmbeddingCalculator.php
@@ -10,6 +10,10 @@ use PhpLlm\LlmChain\Document\Vector;
 use function array_map;
 use function assert;
 use function is_array;
+use function strlen;
+use function substr;
+use function Symfony\Component\String\u;
+use function trim;
 
 class EmbeddingCalculator
 {
@@ -49,5 +53,30 @@ class EmbeddingCalculator
         );
 
         return $embeddings;
+    }
+
+    /** @return list<string> */
+    public function createTextChunks(string $content, int $chunkLength = 800, int $minChunkLength = 50): array
+    {
+        $contentChunks = [];
+        do {
+            $vectorContent     = u($content)->truncate($chunkLength, '', false)->toString();
+            $leftContent       = trim(substr($content, strlen($vectorContent)));
+            $leftContentLength = strlen($leftContent);
+
+            if ($leftContentLength < $minChunkLength || $leftContentLength < $chunkLength / 2) {
+                // If there is too less content left utilize the full left content and end loop
+                $vectorContent = $content;
+                $leftContent   = '';
+            }
+
+            // Set the content to the full left content
+            $content = $leftContent;
+
+            // Now add the chunk to the array
+            $contentChunks[] = $vectorContent;
+        } while ($content !== '');
+
+        return $contentChunks;
     }
 }